+++ .devcontainer/postCreateCommand.sh
@@ -1,3 +1,0 @@
-npm install
-npm run prepack
-npm link